+++ src/main/java/insane96mcp/runeenchanting/data/provider/RELanguageProvider.java
@@ -32,5 +32,12 @@ public class RELanguageProvider extends LanguageProvider {
         add("unknown_curse", "Unknown Curse");
         add("cursed_info", "+1 Socket");
         add("curse_learned", "You have learned about %s");
+
+        add("advancements.runeenchanting.first_rune.title", "First Rune");
+        add("advancements.runeenchanting.first_rune.description", "Obtain a rune");
+        add("advancements.runeenchanting.rune_removed.title", "Clean Slate");
+        add("advancements.runeenchanting.rune_removed.description", "Remove a rune from an item with a grindstone");
+        add("advancements.runeenchanting.curse_learned.title", "Curse Revealed");
+        add("advancements.runeenchanting.curse_learned.description", "Use a cursed item enough to discover its curse");
     }
 }